Changed method for tangent line calculation
Previously, the tangent lines for the functions are based on two points
marking the end of an interval that is plus-minus a fixed distance
around the point line has to be tangent to. However, this results in a
non-constant line-length since in regions of a steeper gradient the
additional height difference will lengthen the tangent line. Now, the
actual slope at the point the line will be tangent to is also
considered. The two points marking the end are now chosen based on the
cosine of the angle the function rises (which equals the arctan of the
slope). However unclear to me, even if this decreases the problem the
tangent lines are still not of equal length for each slope.